Fixed issue when directive binds faster than attributes being compiled
The coordinates for Seattle is passed to googleMaps via scope model, but the value is undefined when
the googleMaps directive gets linked. The solution I found is to use scope.$watch.

`scope.$watch` will wait/listen for changes in the variable, and execute the function body. To tear
down the listener, you call itself. i.e.

    var watcher = scope.$watch('maps, function() {...});
    watcher();