Removing map in detailed view, using the main map instead
Introduces a host of bugs. I keep treating the directives as classes, and as a workaround,
I'm using the view to pass objects in between scopes. I'm sure this is not the 'angular' way.

Backstory:
Google Maps API wants you to keep the map object readily available. Markers, on the other hand, you
can discard the reference once it is bound to a map object. However, map doesn't have a method to
return all markers that is pinned to the map, so to delete any single marker, you still need to keep
an array of markers readily available.

I created a Google Maps directive, and a Markers subdirective. Through controller (as opposed to
link), I can create methods that the subdirectives can call (i.e. it has the scope to get the maps
object), but outside of the directive, I lose the reference to maps object.. unless I store that
object in scope (via directive attribute).

A little hackish, but I can live with it. BUT, now another problem - I can't call the methods I
created in the Google Maps directive. Meaning, duplicate code if I want to use any of those methods
in the directive controller. AKA, I'm treating the directive as a class....

Sigh.